refactor(ontology): Tabular.{row,column}_index: usize → u32
Tabular cell coordinates were usize while ColumnHeader.column_index
was u32. Joining a cell to its header required a cast at every call
site; producers picked whichever felt natural and round-tripped
through casts to make joins work. usize was inherited from
`Vec::get`'s parameter type, not chosen for the data model — CSV /
XLSX / SQL columns are bounded well below u32::MAX.

u32 everywhere on the wire and the public API. Casts now live where
they belong: at the boundary between u32-typed public coordinates
and Rust's usize-typed Vec indices (inside the CSV handler's
`resolve_row` helper and at the read / redact_at Vec lookups).

Closes #218.

+++ crates/nvisy-engine/src/extraction/tabular.rs
@@ -33,7 +33,7 @@ pub(super) async fn populate_document(envelope: &mut DocumentEnvelope<Tabular>) 
     }
 
     // Group cells by row, preserving column order within each row.
-    let mut rows: BTreeMap<usize, Vec<Tabular>> = BTreeMap::new();
+    let mut rows: BTreeMap<u32, Vec<Tabular>> = BTreeMap::new();
     for located in locations {
         rows.entry(located.location.row_index)
             .or_default()
